+++ application/config/constants.php
@@ -83,3 +83,14 @@ defined('EXIT_USER_INPUT')     OR define('EXIT_USER_INPUT', 7); // invalid user 
 defined('EXIT_DATABASE')       OR define('EXIT_DATABASE', 8); // database error
 defined('EXIT__AUTO_MIN')      OR define('EXIT__AUTO_MIN', 9); // lowest automatically-assigned error code
 defined('EXIT__AUTO_MAX')      OR define('EXIT__AUTO_MAX', 125); // highest automatically-assigned error code
+
+/*
+|--------------------------------------------------------------------------
+| CMS Constants
+|--------------------------------------------------------------------------
+|
+*/
+defined('TYPE_DEFAULT')   OR define('TYPE_DEFAULT', 'default');
+defined('TYPE_DROPDOWN')  OR define('TYPE_DROPDOWN', 'dropdown');
+defined('TYPE_SELF_TAB')  OR define('TYPE_SELF_TAB', '_self');
+defined('TYPE_BLANK_TAB') OR define('TYPE_BLANK_TAB', '_blank');