+++ application/migrations/006_create_table_menu.php
@@ -28,16 +28,14 @@ class Migration_Create_table_menu extends CI_Migration
 				'null' => TRUE
 			),
 			'target' => array(
-				'type' => 'INT',
-				'constraint' => '10',
-				'unsigned' => TRUE,
-				'default' => '1'
+				'type' => 'ENUM("_self","_blank")',
+				'default' => '_self',
+				'null' => FALSE
 			),
 			'type' => array(
-				'type' => 'INT',
-				'constraint' => '10',
-				'unsigned' => TRUE,
-				'default' => '0'
+				'type' => 'ENUM("default","dropdown")',
+				'default' => 'default',
+				'null' => FALSE
 			),
 			'parent' => array(
 				'type' => 'BIGINT',
@@ -50,11 +48,11 @@ class Migration_Create_table_menu extends CI_Migration
 		$this->dbforge->create_table('menu');
 
 		$data = array(
-			array('name' => 'More', 'url' => '#', 'icon' => 'fas fa-bars', 'target' => '1', 'type' => '2', 'parent' => 0),
-			array('name' => 'Changelogs', 'url' => 'changelogs', 'icon' => 'fas fa-scroll', 'target' => '1', 'type' => '1', 'parent' => 1),
-			array('name' => 'PvP', 'url' => 'pvp', 'icon' => 'fas fa-fist-raised', 'target' => '1', 'type' => '1', 'parent' => 1),
-			array('name' => 'Forums', 'url' => 'forum', 'icon' => 'fas fa-comments', 'target' => '1', 'type' => '1', 'parent' => 0),
-			array('name' => 'Store', 'url' => 'store', 'icon' => 'fas fa-store', 'target' => '1', 'type' => '1', 'parent' => 0)
+			array('name' => 'More', 'url' => '#', 'icon' => 'fas fa-bars', 'target' => '_self', 'type' => 'dropdown', 'parent' => 0),
+			array('name' => 'Changelogs', 'url' => 'changelogs', 'icon' => 'fas fa-scroll', 'target' => '_self', 'type' => 'default', 'parent' => 1),
+			array('name' => 'PvP', 'url' => 'pvp', 'icon' => 'fas fa-fist-raised', 'target' => '_self', 'type' => 'default', 'parent' => 1),
+			array('name' => 'Forums', 'url' => 'forum', 'icon' => 'fas fa-comments', 'target' => '_self', 'type' => 'default', 'parent' => 0),
+			array('name' => 'Store', 'url' => 'store', 'icon' => 'fas fa-store', 'target' => '_self', 'type' => 'default', 'parent' => 0)
 		);
 		$this->db->insert_batch('menu', $data);
 	}